+++ CHANGELOG.md
@@ -2,13 +2,25 @@
 
 ## [Unreleased]
 
+## [0.2.1]
+
+### Added
+
+- config option for a personal access token
+
+## [0.2.0]
+
 ### Added
 
 - Issues counts
 - Pull Requests counts
-- Customizable emojis via `repofetch.yml` config file
+- Count of available issues with the `help wanted` label
+- Count of available issues with the `good first issue` label
+- Customizable behavior via `repofetch.yml` config file
 
 ## 0.1.0
 Initial version :tada:
 
-[Unreleased]: https://github.com/spenserblack/repofetch/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/spenserblack/repofetch/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/spenserblack/repofetch/compare/v0.2.0...v0.2.1
+[0.2.0]: https://github.com/spenserblack/repofetch/compare/v0.1.0...v0.2.0